Move api/core/v1beta2/index to util/index
